prevent remaining wrap-around

+++ dataplane/dataplaneconnection.h
@@ -55,8 +55,9 @@ private:
     /**
      * @brief remainingBits used in parallel with the fragment buffer, tells how many bits are
      * waiting to be received
+     * a qint32 is used to be able to go under 0 and prevent a wraparound which would cause segfault
      */
-    QHash<quint32, quint16> remainingBits;
+    QHash<quint32, qint32> remainingBits;
     /**
      * @brief totalSize also used in parallel with the two preceding, is there to prevent buffer overflow
      */
